# Bugfix/improvement: Orderbook/ticker update processing