![image](https://i.imgur.com/kPBiGkP.gif)
> Captured footage of GoCryptoTrader attempting to outrun Coinbase Pro updates 

# Description
There have been many updates to websockets, orderbooks and tickers recently that rely on the use of maps. Every time a map is used, go has to look it up - you'll see in the old code, we were looking up items many lines in a row. At the end of our maps are pointers, so we can just make a var referencing the pointer without needing to do a double-lookup.

Coinbase Pro sends out websocket orderbook and ticker updates for _every_ change. When processing this large amount of data, GoCryptoTrader would fall behind leading to an ever larger queue of updates. This was captured by @shazbert as a result of testing https://github.com/thrasher-corp/gocryptotrader/pull/360 (noscope)

This PR addresses that by minimising map lookups, greatly increasing efficiency. Check out the following benchmarks:
*Before*
```
BenchmarkUpdateBidsByPrice-8              	     242	   4810067 ns/op
BenchmarkUpdateAsksByPrice-8              	     254	   4901673 ns/op
BenchmarkBufferPerformance-8              	  131808	     10007 ns/op
BenchmarkBufferSortingPerformance-8       	  117646	     10404 ns/op
BenchmarkBufferSortingByIDPerformance-8   	  115381	     10201 ns/op
BenchmarkNoBufferPerformance-8            	  118810	      9999 ns/op
```
*After*
```
BenchmarkUpdateBidsByPrice-8              	   10000	    115699 ns/op
BenchmarkUpdateAsksByPrice-8              	   10000	    115897 ns/op
BenchmarkBufferPerformance-8              	 1000000	      1245 ns/op
BenchmarkBufferSortingPerformance-8       	  920434	      1298 ns/op
BenchmarkBufferSortingByIDPerformance-8   	  920463	      1310 ns/op
BenchmarkNoBufferPerformance-8            	 1243474	       947 ns/op
```
Now, even when enabling all coinbasepro currencies, disconnecting shows that there is no longer a queue of updates to be processed.

This PR also Fixes the following data race
https://travis-ci.com/thrasher-corp/gocryptotrader/jobs/241323694#L340


## Type of change

Please delete options that are not relevant and add an `x` in `[]` as item is complete.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Improvements to existing code

# How Has This Been Tested?
I wrote the above benchmarks, and `go test ./... -race` and also tested by disconnecting the interenet connection and monitoring whether GCT lagged behind in orderbook updates.

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation and regenerated documentation via the documentation tool 
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally and on Travis with my changes
- [x] Any dependent changes have been merged and published in downstream modules